Move ImplicitByRefModifierInspectionTests for QuickFixes into relevant files under RD.QuickFixes
