+++ RubberduckTests/QuickFixes/IgnoreOnceQuickFixTests.cs
@@ -305,5 +305,33 @@ End Sub";
             Assert.AreEqual(expectedCode, parser.State.GetRewriter(component).GetText());
         }
 
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void ImplicitByRefModifier_IgnoredQuickFixWorks()
+        {
+            const string inputCode =
+@"Sub Foo(arg1 As Integer)
+End Sub";
+
+            const string expectedCode =
+@"'@Ignore ImplicitByRefModifier
+Sub Foo(arg1 As Integer)
+End Sub";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new ImplicitByRefModifierInspection(state);
+            var inspector = InspectionsHelper.GetInspector(inspection);
+            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
+
+            new IgnoreOnceQuickFix(state, new[] { inspection }).Fix(inspectionResults.First());
+
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
+
     }
 }
